Add check for user being logged in

+++ src/controller/Admin.js
@@ -31,51 +31,54 @@ class Admin {
             }
 
         } else {
-            res.render("login", {callback:"/Admin"});
+            res.render("login", {callback:"/Admin"}); //{callback:`/Admin/${page}`});
         }
     }
 
     async post(req, res) {
-        let action = req.params.Action;
-        let objectType = req.params.Page;
-        let resumeDao = require("../dao/ResumeDao");
-        let body = req.body;
-        console.log(body);
+        if(req.session.user.roles.includes("Admin")) {
+            let action = req.params.Action;
+            let objectType = req.params.Page;
+            let resumeDao = require("../dao/ResumeDao");
+            let body = req.body;
 
-        try
-        {
-            switch(action) {
-            case "Delete":
-                if(objectType == "Job") {
-                    resumeDao.deleteJob(body.id);
-                } else if(objectType == "School") {
-                    resumeDao.deleteSchool(body.id);
+            try
+            {
+                switch(action) {
+                case "Delete":
+                    if(objectType == "Job") {
+                        resumeDao.deleteJob(body.id);
+                    } else if(objectType == "School") {
+                        resumeDao.deleteSchool(body.id);
+                    }
+                    break;
+                case "Update":
+                    if(objectType == "Job") {
+                        resumeDao.updateJob(body.id, body.title, body.location, body.startDate, body.endDate, body.description);
+                    } else if(objectType == "School") {
+                        resumeDao.updateSchool(body.id, body.name, body.degree, body.gpa, body.graduated, body.comments);
+                    }
+                    break;
+                case "New":
+                    if(objectType == "Job") {
+                        resumeDao.insertJob(body.title, body.location, body.startDate, body.endDate, body.description);
+                    } else if(objectType == "School") {
+                        resumeDao.insertSchool(body.name, body.degree, body.gpa, body.graduated, body.comments);
+                    }
+                    break;
+                default:
+                    res.redirect(`/Admin/${objectType}`);
+                    break;
                 }
-                break;
-            case "Update":
-                if(objectType == "Job") {
-                    resumeDao.updateJob(body.id, body.title, body.location, body.startDate, body.endDate, body.description);
-                } else if(objectType == "School") {
-                    resumeDao.updateSchool(body.id, body.name, body.degree, body.gpa, body.graduated, body.comments);
-                }
-                break;
-            case "New":
-                if(objectType == "Job") {
-                    resumeDao.insertJob(body.title, body.location, body.startDate, body.endDate, body.description);
-                } else if(objectType == "School") {
-                    resumeDao.insertSchool(body.name, body.degree, body.gpa, body.graduated, body.comments);
-                }
-                break;
-            default:
-                res.redirect(`/Admin/${objectType}`);
-                break;
+            } catch(error) {
+                console.error(error);
+                res.status(500).send(JSON.stringify(error)).end();
             }
-        } catch(error) {
-            console.error(error);
-            res.status(500).send(JSON.stringify(error)).end();
-        }
 
-        res.status(200).send("Sucess").end();
+            res.status(200).send("Sucess").end();
+        }
+    } else {
+        res.redirect('/Error')
     }
 }
